// src/app/generated/model/projectSummaryResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProjectSummaryResponse { 
    /**
     * Identifier of the project this summary is for
     */
    projectId?: number;
    /**
     * Sum of all LABOR-category expenses
     */
    totalLabor?: number;
    /**
     * Sum of all MATERIAL-category expenses
     */
    totalMaterial?: number;
    /**
     * Sum of all OTHER-category expenses
     */
    totalOther?: number;
    /**
     * Sum of every expense regardless of category
     */
    totalExpenses?: number;
    /**
     * The project\'s planned budget, if one was set
     */
    budget?: number;
    /**
     * budget minus totalExpenses; null when the project has no budget set
     */
    budgetRemaining?: number;
}