// src/app/generated/model/projectUpdateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload to replace a project\'s editable fields
 */
export interface ProjectUpdateRequest { 
    /**
     * Project name
     */
    name: string;
    /**
     * Optional free-text description; omit or send null to clear it
     */
    description?: string;
    /**
     * Optional planned budget; omit or send null to clear it
     */
    budget?: number;
    /**
     * Date the project starts
     */
    startDate: string;
    /**
     * Optional planned/actual end date; omit or send null to clear it
     */
    endDate?: string;
}